SK:
-BUG: fixed all warnings in GCC. Only 1 warning remains in the ANN lib.
And several warnings/errors in the BendingEnergyPenalty.
Most warnings were caused by:
1. "Lazy commenting". Example:
  /** this function does something and is not needed *
  void Do() {}
  
  /** this function is needed: */
  void Do2() {}
  
GCC complains about the /** within the /** */ block.
A better way to comment the function away is thus:
  /** this function does something and is not needed */
  void Do() {}
 */
  
  /** this function is needed: */
  void Do2() {}

Or, simply by // before each line.
The lazy commenting does not give errors, but the many warnings are
annoying and clutter the output.

2. comparison of signed and unsigned integers.

3. unused variables.

Some of the signed/unsigned complains actually indicated small bugs in the
code.

+++ src/Components/Optimizers/QuasiNewtonLBFGS/itkQuasiNewtonLBFGSOptimizer.cxx
@@ -286,7 +286,7 @@ namespace itk
 
     searchDir = - gradient;
 
-    unsigned int cp = this->m_Point;
+    int cp = static_cast<int>( this->m_Point );
 
     for (unsigned int i = 0; i < this->m_Bound; ++i)
     {
@@ -321,7 +321,7 @@ namespace itk
         searchDir[j] += alpha_min_beta * s[j];
       }
       ++cp;
-      if ( cp == this->GetMemory() )
+      if ( static_cast<unsigned int>(cp) == this->GetMemory() )
       {
         cp = 0;
       }
